renamed files to kebab-case and upadted linting rules to enforce it
