+++ apps/signup-form/.storybook/preview.tsx
@@ -3,7 +3,7 @@ import i18nLib from '@tryghost/i18n';
 
 import type {Preview} from "@storybook/react";
 import './storybook.css';
-import {AppContextProvider, AppContextType} from '../src/AppContext';
+import {AppContextProvider, AppContextType} from '../src/app-context';
 
 const transparencyGrid = `url("data:image/svg+xml,%3Csvg width='24' height='24' viewBox='0 0 24 24' xmlns='http://www.w3.org/2000/svg'%3E%3Ctitle%3ERectangle%3C/title%3E%3Cg fill='none' fill-rule='evenodd'%3E%3Cpath fill='%23F2F6F8' d='M0 0h24v24H0z'/%3E%3Cpath fill='%23E5ECF0' d='M0 0h12v12H0zM12 12h12v12H12z'/%3E%3C/g%3E%3C/svg%3E")`
