+++ apps/signup-form/src/app.tsx
@@ -1,9 +1,9 @@
 import React, {ComponentProps} from 'react';
 import i18nLib from '@tryghost/i18n';
 import pages, {Page, PageName} from './pages';
-import {AppContextProvider, AppContextType} from './AppContext';
-import {ContentBox} from './components/ContentBox';
-import {Frame} from './components/Frame';
+import {AppContextProvider, AppContextType} from './app-context';
+import {ContentBox} from './components/content-box';
+import {Frame} from './components/frame';
 import {setupGhostApi} from './utils/api';
 import {useOptions} from './utils/options';
