+++ apps/signup-form/src/components/content-box.tsx
