+++ apps/signup-form/src/components/pages/form-page.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import {FormView} from './FormView';
+import {FormView} from './form-view';
 import {isMinimal} from '../../utils/helpers';
 import {isValidEmail} from '../../utils/validator';
-import {useAppContext} from '../../AppContext';
+import {useAppContext} from '../../app-context';
 
 export const FormPage: React.FC = () => {
     const [error, setError] = React.useState('');
