+++ apps/signup-form/src/components/pages/form-view.tsx
@@ -1,6 +1,6 @@
 import React, {FormEventHandler} from 'react';
 import {ReactComponent as LoadingIcon} from '../../../assets/icons/spinner.svg';
-import {useAppContext} from '../../AppContext';
+import {useAppContext} from '../../app-context';
 
 export const FormView: React.FC<FormProps & {
     isMinimal: boolean
