+++ apps/signup-form/src/utils/helpers.tsx
@@ -1,4 +1,4 @@
-import {SignupFormOptions} from '../AppContext';
+import {SignupFormOptions} from '../app-context';
 
 export type URLHistory = {
     type?: 'post',
